Change "Valuation_date" to "effective_date" in Val Obj
As discussed in Issue #123, the field name `valuation_date` is vague. Valuation reports traditionally can have a number of different relevant dates - including the issue date / report date (when the report was issued), the effective date (the date from which the valuation can be relied upon), and the expiration date (the date on which the valuation can no longer be relied upon - i.e. it's "stale"). Per the TWG discussion, rather than *add* all of these fields, we decided what we really wanted to do was just capture the date on which the valuation is actually effective (leaving that determination up to the implementing system). We don't care about the date of the report or other things that one *might* think we mean by "valuation_date".