Change License to Reflect JSON Schema $comment
As discussed, the only way to drop a notice into the JSON Schema files is to use a `$comment` keyword / field. I updated the license to reflect that the copyright notices should be different in a JSON Schema file vs other "OCF Files" (basically all documentation that's not a JSON Schema file.