Only support current version and samples in OCFVersionType
Prior to this change, `OCFVersionType` maintained a running list of
current and past releases. This was misleading, because it:

a) implied a level of compatibility, both forward and backward, that
   is not guaranteed, and

b) implied that the value of this field did not need to match the
   version one is intending to communicate

We discussed making the version field a `const` and placing the value
directly in the `OCFManifestFile` schema. This change, however,
proposes that we maintain an enum with two values:

1. The current semantic version, and
2. A "sample" version for denoting files created for documentation
   purposes

This allows us to maintain validated samples without updating them on
every version update.